# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "KnowledgeVerse API",
    "version": "1.0.0",
    "description": "API Documentation for KnowledgeVerse Backend"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/learning/content": {
      "post": {
        "summary": "Fetch Random Questions for a Subtopic or Topic",
        "tags": [
          "Learning"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "subtopic_id": {
                    "type": "string",
                    "example": "c0000009-0001-0000-0000-000000000065"
                  },
                  "topic_id": {
                    "type": "string",
                    "example": "c0000001-0001-0000-0000-000000000000"
                  },
                  "limit": {
                    "type": "integer",
                    "example": 5
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response containing random questions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "questions": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to retrieve learning content"
          }
        }
      }
    },
    "/api/learning/chat": {
      "post": {
        "summary": "Chat with AI Tutor using Groq",
        "tags": [
          "Learning"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string",
                    "example": "Can you explain Newton's second law?"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "AI Tutor response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "response": {
                      "type": "string",
                      "example": "Newton's second law states that Force equals mass times acceleration (F=ma)."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/learning/submit-quiz": {
      "post": {
        "summary": "Submit Quiz Answers and Award XP & Coins",
        "tags": [
          "Learning"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user_id": {
                    "type": "string"
                  },
                  "building_id": {
                    "type": "string"
                  },
                  "subject": {
                    "type": "string"
                  },
                  "correct_answers": {
                    "type": "integer"
                  },
                  "total_questions": {
                    "type": "integer"
                  },
                  "difficulty": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "500": {
            "description": "Failed to submit quiz results"
          }
        }
      }
    },
    "/api/learning/tts": {
      "post": {
        "summary": "TTS Endpoint for Question / Lesson Audio using ElevenLabs",
        "tags": [
          "Learning"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": {
                    "type": "string",
                    "example": "Welcome to KnowledgeVerse!"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response containing raw audio stream",
            "content": {
              "audio/mpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "500": {
            "description": "Error"
          }
        }
      }
    },
...
  },
...
}